Update script.js
I believe I figured it out. I made the primitive multiply a and b values to 8, and then multiplied them by each other. Then made a catch function to make sure that if there ever was an error a message was sent. I also tied reliable multiply with primitive multiply